+++ frontend/src/components/ui/card.tsx
@@ -1,18 +1,18 @@
 import * as React from "react"
 
-import {cn} from "@/lib/utils"
+import { cn } from "@/lib/utils"
 
 function Card({
-                className,
-                size = "default",
-                ...props
-              }: React.ComponentProps<"div"> & { size?: "default" | "sm" }) {
+  className,
+  size = "default",
+  ...props
+}: React.ComponentProps<"div"> & { size?: "default" | "sm" }) {
   return (
     <div
       data-slot="card"
       data-size={size}
       className={cn(
-        "group/card flex flex-col gap-(--card-spacing) overflow-hidden rounded-xl bg-card py-(--card-spacing) text-sm text-card-foreground ring-1 ring-foreground/10 [--card-spacing:--spacing(4)] has-data-[slot=card-footer]:pb-0 has-[>img:first-child]:pt-0 data-[size=sm]:[--card-spacing:--spacing(3)] data-[size=sm]:has-data-[slot=card-footer]:pb-0 *:[img:first-child]:rounded-t-xl *:[img:last-child]:rounded-b-xl",
+        "group/card flex flex-col gap-(--card-spacing) overflow-hidden rounded-xl bg-card py-(--card-spacing) text-sm text-card-foreground shadow-xs ring-1 ring-foreground/10 [--card-spacing:--spacing(6)] has-[>img:first-child]:pt-0 data-[size=sm]:[--card-spacing:--spacing(4)] *:[img:first-child]:rounded-t-xl *:[img:last-child]:rounded-b-xl",
         className
       )}
       {...props}
@@ -20,7 +20,7 @@ function Card({
   )
 }
 
-function CardHeader({className, ...props}: React.ComponentProps<"div">) {
+function CardHeader({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       data-slot="card-header"
@@ -33,12 +33,12 @@ function CardHeader({className, ...props}: React.ComponentProps<"div">) {
   )
 }
 
-function CardTitle({className, ...props}: React.ComponentProps<"div">) {
+function CardTitle({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       data-slot="card-title"
       className={cn(
-        "cn-font-heading text-base leading-snug font-medium group-data-[size=sm]/card:text-sm",
+        "font-heading text-base leading-normal font-medium group-data-[size=sm]/card:text-sm",
         className
       )}
       {...props}
@@ -46,7 +46,7 @@ function CardTitle({className, ...props}: React.ComponentProps<"div">) {
   )
 }
 
-function CardDescription({className, ...props}: React.ComponentProps<"div">) {
+function CardDescription({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       data-slot="card-description"
@@ -56,7 +56,7 @@ function CardDescription({className, ...props}: React.ComponentProps<"div">) {
   )
 }
 
-function CardAction({className, ...props}: React.ComponentProps<"div">) {
+function CardAction({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       data-slot="card-action"
@@ -69,7 +69,7 @@ function CardAction({className, ...props}: React.ComponentProps<"div">) {
   )
 }
 
-function CardContent({className, ...props}: React.ComponentProps<"div">) {
+function CardContent({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       data-slot="card-content"
@@ -79,12 +79,12 @@ function CardContent({className, ...props}: React.ComponentProps<"div">) {
   )
 }
 
-function CardFooter({className, ...props}: React.ComponentProps<"div">) {
+function CardFooter({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       data-slot="card-footer"
       className={cn(
-        "flex items-center rounded-b-xl border-t bg-muted/50 p-(--card-spacing)",
+        "flex items-center rounded-b-xl px-(--card-spacing) [.border-t]:pt-(--card-spacing)",
         className
       )}
       {...props}